requirements.txt
anthropic>=0.40.0
psycopg[binary]>=3.1.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
logtail-python>=0.3.0

